split42: fix off-screen status text (signed glyph offsets)
The portrait status renderer read glyph xOffset/yOffset with
`int xo = pgm_read_byte(...)`, but pgm_read_byte zero-extends the
Adafruit-GFX int8_t offsets — a text glyph's yOffset of -8 became 248, so
every string and icon plotted far off-screen and was silently clipped by
pset() (only the raw USB/Link bitmaps, the globe and the brightness bars
still drew). Read the offsets through int8_t in both draw paths, matching
disp_array.c. Width/height/xAdvance are always positive and unchanged.

The Python preview parses signed decimals directly, so it validated the
layout but not the compiled C's sign handling — it never reproduced this.
Its --diag mode now also exercises every short layout name (glyph widths
differ) so a name-specific clip can't hide behind "Qwrty".

Also records both traps (signed-offset read + the font-header
double-definition extern) in the qmk CLAUDE.md status-OLED section.

+++ keyboards/polykybd/split42/status_oled.c
@@ -89,7 +89,7 @@ static int pdraw_glyph(const GFXfont* const* fonts, uint8_t n, int x, int baseli
     if (!g) return 0;
     uint16_t bo = pgm_read_word(&g->bitmapOffset);
     int w = pgm_read_byte(&g->width),  h  = pgm_read_byte(&g->height);
-    int xo = pgm_read_byte(&g->xOffset), yo = pgm_read_byte(&g->yOffset);
+    int xo = (int8_t)pgm_read_byte(&g->xOffset), yo = (int8_t)pgm_read_byte(&g->yOffset);
     const uint8_t* bmp = (const uint8_t*)pgm_read_ptr(&f->bitmap);
     int bit = 0; uint8_t bits = 0;
     for (int gy = 0; gy < h; gy++)
@@ -155,7 +155,7 @@ static void pdraw_text_center_half(const GFXfont* const* fonts, uint8_t n, int t
         if (!g) { cx += 4; continue; }   // fallback advance (e.g. a space with no glyph)
         uint16_t bo = pgm_read_word(&g->bitmapOffset);
         int w = pgm_read_byte(&g->width),  h  = pgm_read_byte(&g->height);
-        int xo = pgm_read_byte(&g->xOffset), yo = pgm_read_byte(&g->yOffset);
+        int xo = (int8_t)pgm_read_byte(&g->xOffset), yo = (int8_t)pgm_read_byte(&g->yOffset);
         const uint8_t* bmp = (const uint8_t*)pgm_read_ptr(&f->bitmap);
         int bit = 0; uint8_t bits = 0;
         for (int gy = 0; gy < h; gy++)
